refactor: split it_this.rs into type_subst, chain, receiver, cst_lambda
2201-line god module split into 4 focused sub-modules:
- type_subst.rs: generic type-parameter substitution helpers
- chain.rs: CST navigation-chain type resolution
- receiver.rs: lambda receiver type inference from text context
- cst_lambda.rs: CST-backed ThisLambdaCtx + lambda context helpers

it_this.rs retains only: public API, scanning loop helpers, LambdaParamKind.

No logic changes — verbatim function body extraction.
